when user addnew commande from wakalni we save a facture and commande in wakalni's DB

+++ Restaurants/feane/src/main/java/chmin9lewis/Restaurants/feane/Metier/CommandeMetier.java
@@ -30,7 +30,7 @@ public class CommandeMetier implements ICommandeMetier{
 	public Order addCommande(Order order) {
 		
 		Collection<Plat> plats =  new ArrayList<Plat>();
-		float T=0;//Totalle
+		double T=0;//Totalle
 		try {
 			plats = order.getPlats();
 			for(Plat p : plats) {
